+ further updates
just how many of them?

+++ Dota2Cheat/SDK/Base/NormalClass.h
@@ -5,14 +5,14 @@
 class NormalClass {
 public:
 	template<typename T>
-	T& Field(int offset) const {
+	T& Field(ptrdiff_t offset) const {
 		if (!IsValidReadPtr((uintptr_t)this + offset))
 			throw "VClass::Field access violation";
 		return *(T*)((uintptr_t)this + offset);
 	}
 
 	template<typename T>
-	T Member(int offset/*, T defaultValue = T{}*/) const {
+	T Member(ptrdiff_t offset/*, T defaultValue = T{}*/) const {
 		if (!IsValidReadPtr((uintptr_t)this + offset))
 			return T{};
 		return *(T*)((uintptr_t)this + offset);
@@ -20,7 +20,7 @@ public:
 
 	// Gets a pointer to a type via the offset but does not dereference it
 	template<typename T>
-	T* MemberInline(int offset) const {
+	T* MemberInline(ptrdiff_t offset) const {
 		return (T*)((uintptr_t)this + offset);
 	}
 };
