+++ Dota2Cheat/SDK/Base/VClass.h
@@ -8,14 +8,14 @@ public:
 	virtual void dummy_fn() = 0; // so that classes have a vtable
 
 	template<typename T>
-	T& Field(int offset) const {
+	T& Field(ptrdiff_t offset) const {
 		if (!IsValidReadPtr((uintptr_t)this + offset))
 			throw "VClass::Field access violation";
 		return *(T*)((uintptr_t)this + offset);
 	}
 
 	template<typename T>
-	T Member(int offset/*, T defaultValue = T{}*/) const {
+	T Member(ptrdiff_t offset/*, T defaultValue = T{}*/) const {
 		if (!IsValidReadPtr((uintptr_t)this + offset))
 			return T{};
 		return *(T*)((uintptr_t)this + offset);
@@ -23,11 +23,11 @@ public:
 
 	// Gets a pointer to a type via the offset but does not dereference it
 	template<typename T>
-	T* MemberInline(int offset) const {
+	T* MemberInline(ptrdiff_t offset) const {
 		return (T*)((uintptr_t)this + offset);
 	}
 
-	Function GetVFunc(int index) const
+	Function GetVFunc(ptrdiff_t index) const
 	{
 		return Function((*(uintptr_t**)this)[index]);
 	}
